+++ src/main/java/com/Apharma/sep4/DAO/iSensorDAO.java
@@ -3,6 +3,6 @@ package com.Apharma.sep4.DAO;
 public interface iSensorDAO
 {
 	
-	void saveConstraints(int sensorId, double minValue, double maxValue);
+	void saveConstraints(int sensorId, int minValue, int maxValue);
 	
 }
